requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20
langchain-tavily>=0.1.0
tavily-python>=0.3.0
wikipedia>=1.4.0
arxiv>=1.4.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
requests>=2.31.0